# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Her Planner API - AI-powered life planner for women
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "201":
          description: Created profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
        - name: category
          in: query
          schema:
            type: string
        - name: completed
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Profile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hasKids:
          type: boolean
        numberOfKids:
          type: integer
          nullable: true
        workSchedule:
          type: string
          nullable: true
        healthConditions:
          type: string
          nullable: true
        averageSleepHours:
          type: number
          nullable: true
        cycleLength:
          type: integer
          nullable: true
        periodLength:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - hasKids
        - createdAt
        - updatedAt
    CreateProfileBody:
      type: object
      properties:
        name:
          type: string
        hasKids:
          type: boolean
        numberOfKids:
          type: integer
          nullable: true
        workSchedule:
          type: string
          nullable: true
        healthConditions:
          type: string
          nullable: true
        averageSleepHours:
          type: number
          nullable: true
        cycleLength:
          type: integer
          nullable: true
        periodLength:
          type: integer
          nullable: true
      required:
        - name
        - hasKids
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        hasKids:
          type: boolean
        numberOfKids:
          type: integer
          nullable: true
        workSchedule:
          type: string
          nullable: true
        healthConditions:
          type: string
          nullable: true
        averageSleepHours:
          type: number
          nullable: true
        cycleLength:
          type: integer
          nullable: true
        periodLength:
          type: integer
          nullable: true
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        completed:
          type: boolean
        dueDate:
          type: string
          format: date
          nullable: true
        view:
          type: string
          enum: [today, week, month]
        aiSuggested:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - priority
        - completed
        - view
        - aiSuggested
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date
          nullable: true
        view:
          type: string
          enum: [today, week, month]
        aiSuggested:
          type: boolean
      required:
        - title
        - category
        - priority
        - view
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        completed:
          type: boolean
        dueDate:
          type: string
          format: date
          nullable: true
        view:
          type: string
          enum: [today, week, month]
...