# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  name: string;
  hasKids: boolean;
  numberOfKids?: number | null;
  workSchedule?: string | null;
  healthConditions?: string | null;
  averageSleepHours?: number | null;
  cycleLength?: number | null;
  periodLength?: number | null;
  createdAt: Date;
  updatedAt: Date;
}