# lib/api-zod/src/generated/types/updateTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Her Planner API - AI-powered life planner for women
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateTaskBodyPriority } from "./updateTaskBodyPriority";
import type { UpdateTaskBodyView } from "./updateTaskBodyView";

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  category?: string;
  priority?: UpdateTaskBodyPriority;
  completed?: boolean;
  dueDate?: Date | null;
  view?: UpdateTaskBodyView;
}